Move code to meminfo
