+++ runtime/chainerx_util.h
@@ -80,9 +80,5 @@ chainerx::Array GroupedConvGradWeight(
 
 chainerx::Array SlowRound(const chainerx::Array& a);
 
-void InitializeMemoryMonitoring(chainerx::Device* device);
-size_t GetPeakMemory();
-size_t GetTotalMemory();
-
 }  // namespace runtime
 }  // namespace chainer_compiler
